Fix #59
